requirements.txt
streamlit
transformers
torch
requests
google-genai
python-dotenv